requirements.txt
httpx
python-dotenv
PyJWT
requests
cryptography
pytest
boto3
boto3-stubs
langchain-anthropic